--- src/test/java/com/branas/domain/services/AccountServiceTest.java
@@ -1,25 +1,0 @@
-package com.branas.domain.services;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-class AccountServiceTest {
-
-    @Test
-    void sendEmail() {
-    }
-
-    @Test
-    void validateCpf() {
-    }
-
-    @Test
-    void signup() {
-    }
-
-    @Test
-    void getAccount() {
-    }
-}
